vkr_tg-bot/application_telegram/requirements.txt
aiobotocore==2.15.2
aiocron==1.8
aiofiles==23.2.1
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.5.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
botocore==1.35.36
certifi==2024.7.4
charset-normalizer==3.2.0
colorama==0.4.6
croniter==2.0.2
frozenlist==1.4.0
fsspec==2024.10.0
greenlet==2.0.2
idna==3.7
jmespath==1.0.1
lxml==4.9.3
magic-filter==1.0.12
Mako==1.3.0
MarkupSafe==2.1.4
multidict==6.0.4
pillow==10.3.0
propcache==0.2.0
pydantic==2.4.1
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2024.1
qrcode==8.0
requests==2.32.0
s3fs==2024.10.0
schedule==1.2.1
six==1.16.0
SQLAlchemy==2.0.20
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.17.1
boto3==1.35.16
fastapi==0.115.5
uvicorn==0.32.1
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.10.1
python-multipart==0.0.19